Various cleanups
Refactors the credential loading/handling just a bit to ensure that
credentials from a single "config load" are bundled together and
stored as a `Credentials` object in the config store. This lets us
"clear" any previous credential individual parts when we do a reload,
which is important because we don't want to use, for example, a stale
access token when doing a role_arn token refresh.

Also added a TempFile CloudTest helper.

Tried to add MD5 content hashing for azure, but azurite seems to not
like our md5 of compressed data for some reason?? If real azure is
fine, then we can just disable the m5 hashing for azurite tests.