Remove use of Scratch.jl for writable testing directories (#12)
* Remove use of Scratch.jl for writable testing directories

Turns out this is more pain than it's worth; it doesn't play nicely
when relocatability is required (i.e. PackageCompiler) and it gets
weird when people start playing with DEPOT_PATH and when the parent
scratch directory gets set as read-only. So just use plain old
`mktempdir` and we'll hope for the best.

* whoops