Cleanup azure sas token credentials and usage
I mistakenly mixed up how access tokens and sas tokens are used for azure
credentials. Access tokens are the result of an oauth flow or retrieved from
and azure vm managed identity. SAS tokens are just the query params w/ a
computed sig based on a shared account key (or service level, or user-level
w/ a user delegate key). So SAS tokens can be used to create a "signed" uri,
but the account-level SAS can be used across uris, so it makes sense to have
a dedicated AzureAuth for it.