Make usage of Azure.SASToken idempotent
The issue here was how we were including the SASToken for authorization
in the request url. We were adding it to the url query, regardless
of what was already present. The problem is in the case of retries,
we ended up adding the SASToken multiple times to the query.
This parses the url query and sas token and ensures the individual
token pieces are only present once in the query before reserializing
the request url/target.